+++ src/widgets/UnitEditor/UnitEditor.tsx
@@ -1,15 +1,15 @@
-import { useState } from "react";
-import { calcUnitCost, prepareUnitData } from "../../logic";
+import { useReducer, useState } from "react";
+import { calcUnitCost, formatDistance, isUnitArtilery, isWeaponArtillery, isWeaponMelee, isWeaponRanged, prepareUnitData } from "../../logic";
 import { BasicForm } from "../BasicForm";
 import { FormField } from "../FormField";
-import { WeaponType, weaponsTypes } from "../../data/weapons";
-import { UnitDescriptor, unitTypes } from "../../data/units";
-import { StatsColumns } from "../StatsColumns";
+import { WeaponType, weaponsRange, weaponsTypes } from "../../data/weapons";
+import { UnitDescriptor, unitMovement, unitTypes } from "../../data/units";
 import { ShootingField } from "../ShootingField";
 import { SpecialsField } from "../SpecialsField";
 import { WrittenField } from "../WrittenField";
 import { useTranslation } from "react-i18next";
 import { useAffiliations } from "../../state/useAffiliations";
+import { PointsBadge } from "../PointsBadge";
 
 import "./UnitEditor.css";
 
@@ -20,13 +20,17 @@ export interface UnitEditorProps {
     onCancel?: () => void;
 };
 
+function composeUnit(state: UnitDescriptor, action: Partial<UnitDescriptor>) : UnitDescriptor {
+    return prepareUnitData({ ...state, ...action });
+}
+
 export function UnitEditor({ unit, onSubmit, onCancel } : UnitEditorProps) {
 
     const { t } = useTranslation();
+    const affiliations = useAffiliations();
 
     const [ weapon, setWeapon ] = useState<WeaponType>("pistols");
-    const [ points, setPoints ] = useState<number>(unit ? calcUnitCost(unit) : 0);
-    const affiliations = useAffiliations();
+    const [ currentUnit, changeUnit ] = useReducer(composeUnit, { }, () => unit || prepareUnitData({ })); 
 
     const handleSubmit = (data: object) => {
 
@@ -35,8 +39,7 @@ export function UnitEditor({ unit, onSubmit, onCancel } : UnitEditorProps) {
 
     const handleChange = (data: object) => {
 
-        const unitData = prepareUnitData(data);
-        setPoints(calcUnitCost(unitData));
+        changeUnit(data);
     };
 
     const handleWeaponChange = (value: string) => {
@@ -47,27 +50,92 @@ export function UnitEditor({ unit, onSubmit, onCancel } : UnitEditorProps) {
         onCancel?.();
     };
 
+    const typeDescription = (() => {
+
+        return t(`uniteditor.type.${currentUnit.type}.description`, { distance: formatDistance(unitMovement[currentUnit.type])});
+    })();
+
+    const arnamentDescription = (() => {
+        if (isWeaponArtillery(currentUnit.arnament)) return t("uniteditor.arnament.artillery.description", { distance: formatDistance(weaponsRange[currentUnit.arnament as keyof typeof weaponsRange]) });
+        if (isWeaponRanged(currentUnit.arnament)) return t("uniteditor.arnament.shooting.description", { distance: formatDistance(weaponsRange[currentUnit.arnament as keyof typeof weaponsRange]) });
+        return t("uniteditor.arnament.melee.description");
+    })();
+
     return (
         <BasicForm onChange={handleChange} onSubmit={handleSubmit}>
             <datalist id="uniteditor-affiliation-suggestions">
                 {affiliations.map(a => (<option key={a} value={a}/>))}
             </datalist>
             <div className="uniteditor-firstline">
-                <WrittenField name="name" placeholder="Unit name..." defaultValue={unit?.name}/>
-                <WrittenField name="affiliation" placeholder="Affiliated by..." list="uniteditor-affiliation-suggestions" defaultValue={unit?.affiliation}/>
-                <span>{points} points</span>
-                {onCancel && <button type="button" onClick={handleCancel}>Cancel</button>}
-                <button>Save</button>
+                <WrittenField
+                    name="name"
+                    placeholder={String(t("uniteditor.name.placeholder"))}
+                    defaultValue={currentUnit?.name}
+                />
+                <PointsBadge points={calcUnitCost(currentUnit)} layout="column"/> 
+                {onCancel && <button type="button" onClick={handleCancel}>{t("uniteditor.cancel.label")}</button>}
+                {onSubmit && <button>{t("uniteditor.save.label")}</button>}
             </div>
-            <StatsColumns>
-                <FormField label={t("uniteditor.label.type")} type="select" name="type" options={unitTypes} labels={(o: string) => t(`unit-type.label.${o}`)} defaultValue={unit?.type}/>
-                <FormField label={t("uniteditor.label.arnament")} type="select" name="arnament" options={weaponsTypes} labels={(o: string) => t(`weapon.label.${o}`)} onChange={handleWeaponChange} defaultValue={unit?.arnament}/>
-                <FormField label={t("uniteditor.label.hand-to-hand")} name="handToHand" type="number" min="1" defaultValue={unit?.handToHand}/>
-                <ShootingField label={t("uniteditor.label.shooting")} name="shooting" weapon={weapon} defaultValue={unit?.shooting}/>
-                <FormField label={t("uniteditor.label.morale")} name="morale" type="number" min="1" defaultValue={unit?.morale}/>
-                <FormField label={t("uniteditor.label.stamina")} name="stamina" type="number" min="1" defaultValue={unit?.stamina}/>
-            </StatsColumns>
-            <SpecialsField label={t("uniteditor.label.special")} name="special" defaultValue={unit?.special} />
+            <div className="uniteditor-secondline">
+                <FormField
+                    label={t("uniteditor.affiliation.label")}
+                    type="text"
+                    name="affiliation"
+                    list="uniteditor-affiliation-suggestions"
+                    defaultValue={currentUnit?.affiliation}
+                />
+                <FormField
+                    label={t("uniteditor.type.label")}
+                    type="select"
+                    name="type"
+                    options={unitTypes}
+                    labels={(o: string) => t(`unit-type.label.${o}`)}
+                    defaultValue={currentUnit?.type}
+                    description={typeDescription}
+                />
+                <FormField
+                    label={t("uniteditor.arnament.label")}
+                    type="select"
+                    name="arnament"
+                    options={weaponsTypes}
+                    labels={(o: string) => t(`weapon.label.${o}`)}
+                    onChange={handleWeaponChange}
+                    defaultValue={currentUnit?.arnament}
+                    description={arnamentDescription}
+                />
+            </div>
+            <div className="uniteditor-thirdline">
+                <FormField
+                    label={t("uniteditor.hand-to-hand.label")}
+                    name="handToHand"
+                    type="number"
+                    min="1"
+                    defaultValue={currentUnit?.handToHand}
+                />
+                <ShootingField
+                    label={t("uniteditor.shooting.label")}
+                    name="shooting"
+                    weapon={weapon}
+                    defaultValue={currentUnit?.shooting}
+                />
+                <FormField
+                    label={t("uniteditor.morale.label")}
+                    name="morale"
+                    type="number"
+                    min="1"
+                    defaultValue={currentUnit?.morale}
+                />
+                <FormField
+                    label={t("uniteditor.stamina.label")}
+                    name="stamina"
+                    type="number"
+                    min="1"
+                    defaultValue={currentUnit?.stamina}
+                />
+            </div>
+            <div className="uniteditor-fourthline">
+                <SpecialsField label={t("uniteditor.special.label")} name="special" defaultValue={currentUnit?.special} />
+            </div>
         </BasicForm>
     );
 };