+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+- Added explanations to staff rating in commander editor.
+- Added explanations to personality input in commander editor.
+- Added explanations to weapons type input in unit editor.
+- Added explanations to unit type input in unit editor.
+- Improved layout of unit editor.
 - Improved layout of commander editor.
 - Localization of staff rating.
 - Localization of commander's personalities.
@@ -9,7 +14,7 @@
 - Added cancel buttons to choose unit dialog and choose commander dialogs.
 - It's possible to remove unit from a brigade.
 
-## [0.1.0] - 25/04/2024
+## [0.1.0-beta] - 25/04/2024
 
 - Added changelog page.
 - Added some British and French default units.
@@ -20,7 +25,7 @@
 - Added setup for localization.
 - Updated README.md
 
-## [0.0.0] - 24/04/2024
+## [0.0.0-alpha] - 24/04/2024
 
 - Added prototype for list management.
 - Added stub for rules.